Moved skybox to own file
